--- src/main/java/com/example/likelion/LikelionApplication.java
@@ -1,13 +1,0 @@
-package com.example.likelion;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class LikelionApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(LikelionApplication.class, args);
-    }
-
-}
