--- src/main/java/com/example/likelion/week2/People.java
@@ -1,11 +1,0 @@
-package com.example.likelion.week2;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class People {
-
-    private final String name;
-    private final int age;
-
-}
