--- src/main/java/com/example/likelion/week2/spring/BeanConfig.java
@@ -1,13 +1,0 @@
-package com.example.likelion.week2.spring;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class BeanConfig {
-
-    @Bean
-    public DefaultService defaultService() {
-        return new DefaultService();
-    }
-}
