--- src/main/java/com/example/likelion/week2/spring/BeanService3.java
@@ -1,15 +1,0 @@
-package com.example.likelion.week2.spring;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class BeanService3 {
-    
-    private final DefaultService defaultService;
-    
-    public void confirm() {
-        System.out.println("defaultService = " + defaultService);
-    }
-}
