--- src/main/java/com/example/likelion/week2/spring/DefaultService.java
@@ -1,4 +1,0 @@
-package com.example.likelion.week2.spring;
-
-public class DefaultService {
-}
