+++ src/main/java/study/likelionbeweekly/LikelionApplication.java
@@ -1,0 +1,13 @@
+package study.likelionbeweekly;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LikelionApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(LikelionApplication.class, args);
+    }
+
+}
