+++ src/main/java/study/likelionbeweekly/week2/People.java
@@ -1,0 +1,11 @@
+package study.likelionbeweekly.week2;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public class People {
+
+    private final String name;
+    private final int age;
+
+}
