+++ src/main/java/study/likelionbeweekly/week2/di/CoffeeManager.java
@@ -1,0 +1,9 @@
+package study.likelionbeweekly.week2.di;
+
+public class CoffeeManager implements Manager {
+
+    @Override
+    public void makeCoffee(){
+        System.out.println("CoffeeManager.makeCoffee");
+    }
+}
