+++ src/main/java/study/likelionbeweekly/week2/di/CoffeeService.java
@@ -1,0 +1,14 @@
+package study.likelionbeweekly.week2.di;
+
+public class CoffeeService {
+
+    private final Manager coffeeManager;
+
+    public CoffeeService(Manager coffeeManager) {
+        this.coffeeManager = coffeeManager;
+    }
+
+    public void sale() {
+        coffeeManager.makeCoffee();
+    }
+}
