+++ src/main/java/study/likelionbeweekly/week2/di/Manager.java
@@ -1,0 +1,8 @@
+package study.likelionbeweekly.week2.di;
+
+public interface Manager {
+
+    void makeCoffee();
+
+}
+
