+++ src/main/java/study/likelionbeweekly/week2/di/NewCoffeeManager.java
@@ -1,0 +1,9 @@
+package study.likelionbeweekly.week2.di;
+
+public class NewCoffeeManager implements Manager {
+
+    @Override
+    public void makeCoffee(){
+        System.out.println("NewCoffeeManager.makeCoffee");
+    }
+}
