+++ src/main/java/study/likelionbeweekly/week2/spring/BeanConfig.java
@@ -1,0 +1,13 @@
+package study.likelionbeweekly.week2.spring;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+public class BeanConfig {
+
+    @Bean
+    public DefaultService defaultService() {
+        return new DefaultService();
+    }
+}
