+++ src/main/java/study/likelionbeweekly/week2/spring/BeanService3.java
@@ -1,0 +1,15 @@
+package study.likelionbeweekly.week2.spring;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class BeanService3 {
+    
+    private final DefaultService defaultService;
+    
+    public void confirm() {
+        System.out.println("defaultService = " + defaultService);
+    }
+}
