+++ src/main/java/study/likelionbeweekly/week2/spring/DefaultService.java
@@ -1,0 +1,4 @@
+package study.likelionbeweekly.week2.spring;
+
+public class DefaultService {
+}
