--- src/test/java/com/example/likelion/LikelionApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.likelion;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class LikelionApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
