+++ src/test/java/study/likelionbeweekly/LikelionApplicationTests.java
@@ -1,0 +1,13 @@
+package study.likelionbeweekly;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LikelionApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
